chapter10/requirements.txt
pytest==8.3.5
pytest-docker-tools==3.1.9
pytest-helpers-namespace==2021.12.29
pytest-mock==3.14.0
flaky==3.8.1
apache-airflow-providers-postgres==6.2.0
apache-airflow-providers-fab==2.0.2
